move relative time, status, total to utils
